Change create and edit flow
