+++ app/components/shared/bread-crumb/index.js
@@ -42,6 +42,17 @@ export default class SharedBreadCrumbComponent extends Component {
         ],
       },
       {
+        route: 'association.recognition.edit',
+        crumbs: [
+          { label: 'Verenigingen', link: 'index' },
+          {
+            label: name,
+          },
+          { label: 'Erkenningen', link: 'association.recognition.index' },
+          { label: 'Bewerken' },
+        ],
+      },
+      {
         route: 'association.contact-detail',
         crumbs: [
           { label: 'Verenigingen', link: 'index' },
