+++ app/router.js
@@ -25,7 +25,8 @@ Router.map(function () {
     this.route('representatives');
     this.route('recognition', { path: 'erkenningen' }, function () {
       this.route('index', { path: '/' });
-      this.route('create', { path: 'create/:recognition_id' });
+      this.route('create', { path: '/aanmaken' });
+      this.route('edit', { path: '/bewerk/:recognition_id' });
     });
   });
